+++ lending/src/codatlending/models/operations/download_excel_report.py
@@ -26,7 +26,7 @@ class DownloadExcelReportResponse:
     r"""HTTP response status code for this operation"""
     body: Optional[bytes] = dataclasses.field(default=None)
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""