+++ lending/src/codatlending/models/operations/get_categorized_balance_sheet_statement.py
@@ -29,7 +29,7 @@ class GetCategorizedBalanceSheetStatementResponse:
     enhanced_financial_report: Optional[shared_enhancedfinancialreport.EnhancedFinancialReport] = dataclasses.field(default=None)
     r"""OK"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""