+++ lending/src/codatlending/models/operations/is_aged_creditors_report_available.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
+from ..shared import errormessage as shared_errormessage
 from typing import Optional
 
 
@@ -20,6 +21,8 @@ class IsAgedCreditorsReportAvailableResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""Your API request was not properly authorized."""
     is_aged_creditors_report_available_200_application_json_boolean: Optional[bool] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)